update to v1.2.0-stable

+++ app/views/kelas/index.php
@@ -9,6 +9,10 @@ if (!Session::has('authenticated')) {
     return redirect('../../index.php');
 }
 
+if (Session::get('level') != '2') {
+    return abort(404);
+}
+
 $kelas = Dashboard::readKelas();
 
 reqFile('../templates/header.php');
@@ -19,20 +23,26 @@ reqFile('../templates/sidebar.php');
 <section class="home-section">
     <div class="text">Kelas</div>
     <div class="container">
-        <a href="addKelas.php">Tambah Kelas</a>
+        <a href="addKelas.php"><button class="btn btn-add">+ Tambah Kelas</button></a>
         <table>
             <tr>
+                <th>Kelas</th>
                 <th>Nama Kelas</th>
-                <th>kompetensi Keahlian</th>
+                <th>Kompetensi Keahlian</th>
                 <th>Aksi</th>
             </tr>
             <?php foreach ($kelas as $k) : ?>
                 <tr>
-                    <td><?= $k['nama_kelas'] ?></td>
-                    <td><?= $k['kompetensi_keahlian'] ?></td>
+                    <td><?= $k['kelas']; ?></td>
                     <td>
-                        <a href="updateKelas.php?idkelas=<?= $k['id_kelas'] ?>">Edit</a>
-                        <a href="deleteKelas.php?idkelas=<?= $k['id_kelas'] ?>">Hapus</a>
+                        <a href="../dashboard/search.php?qk=<?= str_replace(' ', '-', strtolower($k['nama_kelas'])); ?>">
+                            <button class="btn btn-on-table"><?= $k['nama_kelas'] ?></button>
+                        </a>
+                    </td>
+                    <td><?= $k['kompetensi_keahlian']; ?></td>
+                    <td>
+                        <a href="updateKelas.php?idkelas=<?= $k['id_kelas'] ?>"><button class="btn btn-action-primary">Edit</button></a>
+                        <a href="deleteKelas.php?idkelas=<?= $k['id_kelas'] ?>"><button class="btn btn-action-secondary" onclick="return confirm('Anda yakin?')">Hapus</button></a>
                     </td>
                 </tr>
             <?php endforeach; ?>
